Refactor Slab logic away from WoodenSlab, added StoneSlab2 and its double
