+++ src/pocketmine/block/StoneSlab.php
@@ -26,7 +26,7 @@ namespace pocketmine\block;
 use pocketmine\item\Item;
 use pocketmine\item\Tool;
 
-class StoneSlab extends WoodenSlab{
+class StoneSlab extends Slab{
 	const STONE = 0;
 	const SANDSTONE = 1;
 	const WOODEN = 2;
@@ -38,7 +38,9 @@ class StoneSlab extends WoodenSlab{
 
 	protected $id = self::STONE_SLAB;
 
-	protected $doubleId = self::DOUBLE_STONE_SLAB;
+	public function getDoubleSlabId() : int{
+		return self::DOUBLE_STONE_SLAB;
+	}
 
 	public function getHardness() : float{
 		return 2;
@@ -68,9 +70,5 @@ class StoneSlab extends WoodenSlab{
 		}
 
 		return [];
-	}
-
-	public function getFuelTime() : int{
-		return 0;
 	}
 }
